add the possibility to simulate sphere
